Enhance channel deduction logic
The most probable channel is not the first one found among those with
no DLC errors, but the one with the fewest DLC errors.

This approach provides a result that is likely to be adequate even in
scenarios where the external control unit sends some command frames
with the wrong DLC (for example, if it is incapable of transmitting
frames with a DLC different from 8).